feat(v1-20): Storybook Excellence — 全覆盖 + 巨石拆分 + 文档 (#1227)
- 补充 rightpanel/settings/shortcuts 3 个缺失目录的 Story 文件
- 拆分 9 个巨石 Story（AiPanel/QualityGatesPanel/CharacterPanel/VersionHistory/FileTree/CommandPalette/Search/Outline/DiffView）为 21 个文件，每个 ≤500 行
- 新增 17 个 play function（Dashboard/Settings/Editor/Export/Analytics 等）
- 5 个核心 Primitives（Button/Card/Input/Badge/Tabs）添加使用指南文档
- 双主题 decorator：Storybook 工具栏支持 Dark/Light 切换
- storybook-excellence-guard 测试：覆盖率 + 行数门禁
- 更新 FileTreePanel.storybook-coverage 与 AiPanel.stories.test 适配拆分后文件

Closes #1227
