+++ apps/desktop/renderer/src/features/commandPalette/CommandPaletteAdvanced.stories.tsx
@@ -21,10 +21,6 @@ import {
   createEditorStore,
 } from "../../stores/editorStore";
 
-// =============================================================================
-// Mock IPC for Storybook
-// =============================================================================
-
 function createMockIpc() {
   return {
     invoke: async (channel: string): Promise<unknown> => {
@@ -50,7 +46,6 @@ function createMockIpc() {
   };
 }
 
-// Decorator to provide stores
 function StoreDecorator({ children }: { children: React.ReactNode }) {
   const mockIpc = React.useMemo(() => createMockIpc(), []);
   const projectStore = React.useMemo(
@@ -69,64 +64,8 @@ function StoreDecorator({ children }: { children: React.ReactNode }) {
   );
 }
 
-/**
- * CommandPalette 组件 Story
- *
- * 设计稿参考：design/Variant/designs/17-command-palette.html
- *
- * 功能：
- * - 三段式布局：Header（搜索框）+ Body（分组列表）+ Footer（键盘提示）
- * - 搜索过滤：实时过滤命令/文件
- * - 键盘导航：↑↓ 移动，Enter 确认，Esc 关闭
- * - 搜索高亮：匹配文字高亮显示
- */
-const meta: Meta<typeof CommandPalette> = {
-  title: "Features/CommandPalette",
-  component: CommandPalette,
-  parameters: {
-    layout: "centered",
-  },
-  tags: ["autodocs"],
-  decorators: [
-    (Story) => (
-      <StoreDecorator>
-        <Story />
-      </StoreDecorator>
-    ),
-  ],
-  argTypes: {
-    open: {
-      control: "boolean",
-      description: "Whether the palette is open",
-    },
-  },
-  args: {
-    onOpenChange: fn(),
-  },
-};
-
-export default meta;
-type Story = StoryObj<typeof CommandPalette>;
-
-// =============================================================================
-// Icons for stories (using Lucide)
-// =============================================================================
-
-function FileIcon({
-  className,
-  color,
-}: {
-  className?: string;
-  color?: string;
-}): JSX.Element {
-  return (
-    <FileText
-      className={className}
-      style={{ color }}
-      size={16}
-      strokeWidth={1.5}
-    />
-  );
+function FileIcon({ className, color }: { className?: string; color?: string }): JSX.Element {
+  return <FileText className={className} style={{ color }} size={16} strokeWidth={1.5} />;
 }
 
 function EditIcon({ className }: { className?: string }): JSX.Element {
@@ -152,10 +91,6 @@ function TerminalIcon({ className }: { className?: string }): JSX.Element {
 function DownloadIcon({ className }: { className?: string }): JSX.Element {
   return <Download className={className} size={16} strokeWidth={1.5} />;
 }
-
-// =============================================================================
-// Mock commands for stories
-// =============================================================================
 
 const recentFiles: CommandItem[] = [
   {
@@ -258,201 +193,41 @@ const searchCommands: CommandItem[] = [
   },
 ];
 
-const defaultCommands = [...recentFiles, ...suggestions];
-
-// =============================================================================
-// Stories
-// =============================================================================
-
-/**
- * 默认状态
- *
- * 显示 Recent Files 和 Suggestions 分组
- */
-export const Default: Story = {
-  args: {
-    open: true,
-    commands: defaultCommands,
+const meta: Meta<typeof CommandPalette> = {
+  title: "Features/CommandPalette/Advanced",
+  component: CommandPalette,
+  parameters: {
+    layout: "centered",
   },
-  render: (args) => (
-    <div
-      style={{
-        width: "800px",
-        height: "500px",
-        position: "relative",
-        backgroundColor: "var(--color-bg-base)",
-      }}
-    >
-      <CommandPalette {...args} />
-    </div>
-  ),
-};
-
-/**
- * 搜索状态
- *
- * 搜索 "set" 时的过滤结果，显示匹配高亮
- */
-export const Searching: Story = {
-  args: {
-    open: true,
-    commands: searchCommands,
+  tags: ["autodocs"],
+  decorators: [
+    (Story) => (
+      <StoreDecorator>
+        <Story />
+      </StoreDecorator>
+    ),
+  ],
+  argTypes: {
+    open: {
+      control: "boolean",
+      description: "Whether the palette is open",
+    },
   },
-  render: function SearchingStory(args) {
-    const [open, setOpen] = React.useState(true);
-
-    return (
-      <div
-        style={{
-          width: "800px",
-          height: "500px",
-          position: "relative",
-          backgroundColor: "var(--color-bg-base)",
-        }}
-      >
-        <CommandPalette {...args} open={open} onOpenChange={setOpen} />
-        <div
-          style={{
-            position: "absolute",
-            bottom: "20px",
-            left: "20px",
-            color: "var(--color-fg-muted)",
-            fontSize: "12px",
-          }}
-        >
-          Try typing &quot;set&quot; in the search box to see filtering
-        </div>
-      </div>
-    );
+  args: {
+    onOpenChange: fn(),
   },
 };
 
-/**
- * 空结果状态
- *
- * 搜索无匹配时的空状态
- */
-export const EmptyResults: Story = {
-  args: {
-    open: true,
-    commands: [], // 空命令列表模拟无结果
-  },
-  render: (args) => (
-    <div
-      style={{
-        width: "800px",
-        height: "500px",
-        position: "relative",
-        backgroundColor: "var(--color-bg-base)",
-      }}
-    >
-      <CommandPalette {...args} />
-    </div>
-  ),
+export default meta;
+type Story = StoryObj<typeof CommandPalette>;
+
+const kbdStyle: React.CSSProperties = {
+  backgroundColor: "var(--color-bg-raised)",
+  padding: "2px 4px",
+  borderRadius: "3px",
 };
 
-/**
- * 关闭状态
- *
- * 命令面板关闭（不渲染）
- */
-export const Closed: Story = {
-  args: {
-    open: false,
-    commands: defaultCommands,
-  },
-  render: (args) => (
-    <div
-      style={{
-        width: "800px",
-        height: "300px",
-        position: "relative",
-        backgroundColor: "var(--color-bg-base)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center",
-      }}
-    >
-      <CommandPalette {...args} />
-      <div style={{ color: "var(--color-fg-muted)", textAlign: "center" }}>
-        Command palette is closed (nothing rendered)
-        <br />
-        <span
-          style={{ fontSize: "12px", color: "var(--color-fg-placeholder)" }}
-        >
-          Press Cmd/Ctrl+P to open
-        </span>
-      </div>
-    </div>
-  ),
-};
-
-/**
- * 交互演示
- *
- * 可交互的完整演示（点击按钮打开）
- */
-export const Interactive: Story = {
-  args: {
-    commands: defaultCommands,
-  },
-  render: function InteractiveStory(args) {
-    const [open, setOpen] = React.useState(false);
-
-    React.useEffect(() => {
-      function handleKeyDown(e: KeyboardEvent): void {
-        if ((e.metaKey || e.ctrlKey) && e.key === "p") {
-          e.preventDefault();
-          setOpen(true);
-        }
-      }
-
-      document.addEventListener("keydown", handleKeyDown);
-      return () => document.removeEventListener("keydown", handleKeyDown);
-    }, []);
-
-    return (
-      <div
-        style={{
-          width: "800px",
-          height: "500px",
-          position: "relative",
-          backgroundColor: "var(--color-bg-base)",
-          display: "flex",
-          flexDirection: "column",
-          alignItems: "center",
-          justifyContent: "center",
-          gap: "16px",
-        }}
-      >
-        <button
-          onClick={() => setOpen(true)}
-          style={{
-            padding: "8px 16px",
-            backgroundColor: "var(--color-bg-raised)",
-            border: "1px solid var(--color-border-default)",
-            borderRadius: "var(--radius-sm)",
-            color: "var(--color-fg-default)",
-            cursor: "pointer",
-            fontSize: "13px",
-          }}
-        >
-          Open Command Palette
-        </button>
-        <div style={{ color: "var(--color-fg-placeholder)", fontSize: "12px" }}>
-          or press Cmd/Ctrl+P
-        </div>
-        <CommandPalette {...args} open={open} onOpenChange={setOpen} />
-      </div>
-    );
-  },
-};
-
-/**
- * 多分组
- *
- * 显示多个分组的场景
- */
+/** 多分组 — 显示多个分组的场景 */
 export const MultipleGroups: Story = {
   args: {
     open: true,
@@ -510,29 +285,7 @@ export const MultipleGroups: Story = {
   ),
 };
 
-// =============================================================================
-// P2: 键盘导航测试
-// =============================================================================
-
-/**
- * 键盘导航演示
- *
- * 展示如何使用键盘操作命令面板。
- *
- * 验证点：
- * - ↑↓ 键移动选中项
- * - 选中项有视觉高亮（背景色变化）
- * - Enter 键执行当前选中命令
- * - Esc 键关闭面板
- * - Tab 键在搜索框和列表之间切换焦点
- *
- * 浏览器测试步骤：
- * 1. 按 ↓ 键，验证第一项被选中（高亮）
- * 2. 继续按 ↓，验证高亮移动到下一项
- * 3. 按 ↑ 键，验证高亮返回上一项
- * 4. 按 Enter 键，验证 Actions 面板显示选中的命令
- * 5. 按 Esc 键，验证面板关闭
- */
+/** 键盘导航演示 — ↑↓ 移动选中项，Enter 执行，Esc 关闭 */
 export const KeyboardNavigationDemo: Story = {
   args: {
     open: true,
@@ -596,7 +349,6 @@ export const KeyboardNavigationDemo: Story = {
           backgroundColor: "var(--color-bg-base)",
         }}
       >
-        {/* 操作提示 */}
         <div
           style={{
             position: "absolute",
@@ -615,42 +367,9 @@ export const KeyboardNavigationDemo: Story = {
             键盘导航测试（Windows）：
           </p>
           <ul style={{ paddingLeft: "1rem", margin: 0, lineHeight: 1.6 }}>
-            <li>
-              <code
-                style={{
-                  backgroundColor: "var(--color-bg-raised)",
-                  padding: "2px 4px",
-                  borderRadius: "3px",
-                }}
-              >
-                ↑↓
-              </code>{" "}
-              移动选中项
-            </li>
-            <li>
-              <code
-                style={{
-                  backgroundColor: "var(--color-bg-raised)",
-                  padding: "2px 4px",
-                  borderRadius: "3px",
-                }}
-              >
-                Enter
-              </code>{" "}
-              执行命令
-            </li>
-            <li>
-              <code
-                style={{
-                  backgroundColor: "var(--color-bg-raised)",
-                  padding: "2px 4px",
-                  borderRadius: "3px",
-                }}
-              >
-                Esc
-              </code>{" "}
-              关闭面板
-            </li>
+            <li><code style={kbdStyle}>↑↓</code> 移动选中项</li>
+            <li><code style={kbdStyle}>Enter</code> 执行命令</li>
+            <li><code style={kbdStyle}>Esc</code> 关闭面板</li>
           </ul>
           {lastAction && (
             <div
@@ -666,7 +385,6 @@ export const KeyboardNavigationDemo: Story = {
             </div>
           )}
         </div>
-
         <CommandPalette {...args} />
       </div>
     );
@@ -680,21 +398,7 @@ export const KeyboardNavigationDemo: Story = {
   },
 };
 
-/**
- * 搜索高亮
- *
- * 展示搜索时匹配文字的高亮效果。
- *
- * 验证点：
- * - 搜索 "set" 时，"Settings" 中的 "Set" 高亮
- * - 高亮使用 <mark> 标签
- * - 高亮样式：黄色背景或下划线
- *
- * 浏览器测试步骤：
- * 1. 观察搜索框已有 "set" 文字
- * 2. 验证 "Open Settings" 中 "Set" 部分高亮
- * 3. 验证 "useSettings.ts" 中 "Set" 部分高亮
- */
+/** 搜索高亮 — 展示搜索时匹配文字的高亮效果 */
 export const SearchHighlight: Story = {
   args: {
     open: true,
@@ -736,23 +440,7 @@ export const SearchHighlight: Story = {
   },
 };
 
-/**
- * 长命令列表
- *
- * 测试大量命令时的滚动和分组显示。
- *
- * 验证点：
- * - 50 个命令正常显示
- * - 滚动流畅无卡顿
- * - 分组标题 sticky（固定在顶部）
- * - 键盘导航正常工作
- *
- * 浏览器测试步骤：
- * 1. 滚动列表，验证流畅性
- * 2. 验证分组标题固定在顶部
- * 3. 使用 ↓ 键快速导航到底部
- * 4. 验证滚动自动跟随选中项
- */
+/** 长命令列表 — 测试大量命令时的滚动和分组显示 */
 export const LongCommandList: Story = {
   args: {
     open: true,
