+++ apps/desktop/renderer/src/features/editor/EditorPane.stories.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
+import { within, expect } from "@storybook/test";
 import { EditorContent, useEditor } from "@tiptap/react";
 import StarterKit from "@tiptap/starter-kit";
 import Underline from "@tiptap/extension-underline";
@@ -115,6 +116,11 @@ export const Default: Story = {
       <StandaloneEditor />
     </div>
   ),
+  play: async ({ canvasElement }) => {
+    const canvas = within(canvasElement);
+    await expect(canvas.getByTestId("editor-pane")).toBeInTheDocument();
+    await expect(canvas.getByTestId("editor-toolbar")).toBeInTheDocument();
+  },
 };
 
 /**
@@ -147,6 +153,11 @@ export const WithContent: Story = {
       />
     </div>
   ),
+  play: async ({ canvasElement }) => {
+    const canvas = within(canvasElement);
+    await expect(canvas.getByTestId("editor-pane")).toBeInTheDocument();
+    await expect(canvas.getByTestId("tiptap-editor")).toBeInTheDocument();
+  },
 };
 
 /**
