+++ apps/desktop/renderer/src/features/editor/EditorToolbar.stories.tsx
@@ -1,4 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
+import { within, expect } from "@storybook/test";
 import { useEditor } from "@tiptap/react";
 import StarterKit from "@tiptap/starter-kit";
 import type { CSSProperties } from "react";
@@ -64,6 +65,12 @@ function ToolbarWithEditor(props: { initialContent?: string }): JSX.Element {
  */
 export const Default: Story = {
   render: () => <ToolbarWithEditor />,
+  play: async ({ canvasElement }) => {
+    const canvas = within(canvasElement);
+    await expect(canvas.getByTestId("toolbar-bold")).toBeInTheDocument();
+    await expect(canvas.getByTestId("toolbar-italic")).toBeInTheDocument();
+    await expect(canvas.getByTestId("toolbar-undo")).toBeInTheDocument();
+  },
 };
 
 /**
@@ -92,6 +99,12 @@ export const WithFormattedContent: Story = {
       `}
     />
   ),
+  play: async ({ canvasElement }) => {
+    const canvas = within(canvasElement);
+    await expect(canvas.getByTestId("toolbar-bold")).toBeInTheDocument();
+    await expect(canvas.getByTestId("toolbar-h1")).toBeInTheDocument();
+    await expect(canvas.getByTestId("toolbar-bullet-list")).toBeInTheDocument();
+  },
 };
 
 /**
