+++ apps/desktop/renderer/src/features/outline/OutlineInteractive.stories.tsx
@@ -48,6 +48,21 @@ const SAMPLE_OUTLINE_DATA: OutlineItem[] = [
 ];
 
 /**
+ * Sample word counts for each section
+ */
+const SAMPLE_WORD_COUNTS: Record<string, number> = {
+  "h1-aesthetics": 2450,
+  "h2-intro": 320,
+  "h2-historical": 890,
+  "h3-early": 420,
+  "h3-postwar": 470,
+  "h2-digital": 680,
+  "h3-interface": 680,
+  "h1-conclusion": 560,
+  "h2-future": 560,
+};
+
+/**
  * Flatten nested outline structure for the component
  */
 function flattenOutline(items: OutlineItem[]): OutlineItem[] {
@@ -66,82 +81,20 @@ function flattenOutline(items: OutlineItem[]): OutlineItem[] {
 
 const FLAT_SAMPLE_DATA = flattenOutline(SAMPLE_OUTLINE_DATA);
 
-/**
- * Sample word counts for each section
- */
-const SAMPLE_WORD_COUNTS: Record<string, number> = {
-  "h1-aesthetics": 2450,
-  "h2-intro": 320,
-  "h2-historical": 890,
-  "h3-early": 420,
-  "h3-postwar": 470,
-  "h2-digital": 680,
-  "h3-interface": 680,
-  "h1-conclusion": 560,
-  "h2-future": 560,
-};
-
-/**
- * Extended data with a very long title for truncation testing
- */
-const LONG_TITLE_DATA: OutlineItem[] = [
-  ...FLAT_SAMPLE_DATA.slice(0, 2),
-  {
-    id: "h2-long",
-    title:
-      "This is a very long chapter title that should be truncated with ellipsis when it exceeds the available width",
-    level: "h2",
-  },
-  ...FLAT_SAMPLE_DATA.slice(2),
-];
-
-/**
- * Large document data for performance testing
- */
-function generateLargeOutline(): OutlineItem[] {
-  const items: OutlineItem[] = [];
-  for (let i = 1; i <= 10; i++) {
-    items.push({
-      id: `ch-${i}`,
-      title: `Chapter ${i}: The Journey Continues`,
-      level: "h1",
-    });
-    for (let j = 1; j <= 5; j++) {
-      items.push({
-        id: `ch-${i}-s-${j}`,
-        title: `${i}.${j} Section Title Here`,
-        level: "h2",
-      });
-      for (let k = 1; k <= 3; k++) {
-        items.push({
-          id: `ch-${i}-s-${j}-ss-${k}`,
-          title: `${i}.${j}.${k} Subsection`,
-          level: "h3",
-        });
-      }
-    }
-  }
-  return items;
-}
-
-const LARGE_OUTLINE_DATA = generateLargeOutline();
-
 const meta: Meta<typeof OutlinePanel> = {
-  title: "Features/OutlinePanel",
+  title: "Features/Outline/Interactive",
   component: OutlinePanel,
   parameters: {
     layout: "fullscreen",
     docs: {
       description: {
-        component: `大纲侧边栏 - 用于显示文档结构和快速导航。
+        component: `大纲侧边栏 - 交互功能。
 
 **功能 (P0)**:
 - 单节点展开/折叠
 - 完整拖拽支持 (before/after/into)
-- 编辑器滚动同步接口
 
 **功能 (P1)**:
-- 字数统计显示
 - 搜索/过滤功能
 - 多选批量操作 (Ctrl/Cmd+Click, Shift+Click)
 - 完整键盘导航 (Arrow/F2/Delete)
@@ -197,33 +150,8 @@ export default meta;
 type Story = StoryObj<typeof OutlinePanel>;
 
 // =============================================================================
-// P0 Stories
+// P0 Interactive Stories
 // =============================================================================
-
-/**
- * Scene 1: DefaultMultiLevel
- *
- * 完整多层级大纲，验证基础渲染
- */
-export const DefaultMultiLevel: Story = {
-  args: {
-    items: FLAT_SAMPLE_DATA,
-    activeId: "h1-aesthetics",
-    draggable: true,
-  },
-};
-
-/**
- * Scene 2: EmptyDocument
- *
- * 空文档无大纲
- */
-export const EmptyDocument: Story = {
-  args: {
-    items: [],
-    activeId: null,
-  },
-};
 
 /**
  * Scene 3: SingleNodeCollapse (P0)
@@ -275,7 +203,6 @@ function DragDropCompleteRender() {
     position: DropPosition,
   ) => {
     console.log(`Reorder: ${draggedId} -> ${position} ${targetId}`);
-    // Simple reorder demo (actual implementation would be more complex)
     const newItems = [...items];
     const draggedIndex = newItems.findIndex((i) => i.id === draggedId);
     const targetIndex = newItems.findIndex((i) => i.id === targetId);
@@ -317,53 +244,9 @@ export const DragDropComplete: Story = {
   },
 };
 
-/**
- * Scene 5: EditorScrollSync (P0)
- *
- * 编辑器滚动同步指示器
- */
-export const EditorScrollSync: Story = {
-  args: {
-    items: FLAT_SAMPLE_DATA,
-    activeId: "h2-historical",
-    scrollSyncEnabled: true,
-    draggable: false,
-  },
-  parameters: {
-    docs: {
-      description: {
-        story:
-          "P0: 编辑器滚动同步。启用后底部显示绿色同步指示器，activeId 会随编辑器滚动自动更新。",
-      },
-    },
-  },
-};
-
 // =============================================================================
-// P1 Stories
+// P1 Interactive Stories
 // =============================================================================
-
-/**
- * Scene 6: WordCountDisplay (P1)
- *
- * 字数统计显示
- */
-export const WordCountDisplay: Story = {
-  args: {
-    items: FLAT_SAMPLE_DATA,
-    activeId: "h1-aesthetics",
-    wordCounts: SAMPLE_WORD_COUNTS,
-    draggable: false,
-  },
-  parameters: {
-    docs: {
-      description: {
-        story:
-          "P1: 字数统计显示。每个大纲项右侧显示该章节的字数（如 2.4k、320 等）。",
-      },
-    },
-  },
-};
 
 /**
  * Scene 7: SearchFilter (P1)
@@ -477,40 +360,6 @@ export const KeyboardNavigation: Story = {
 - Delete/Backspace: 删除当前项
 - Escape: 清除多选
 - Ctrl/Cmd + A: 全选`,
-      },
-    },
-  },
-};
-
-/**
- * Scene 10: LongTitleTruncation
- *
- * 超长标题截断
- */
-export const LongTitleTruncation: Story = {
-  args: {
-    items: LONG_TITLE_DATA,
-    activeId: "h1-aesthetics",
-    draggable: true,
-  },
-};
-
-/**
- * Scene 11: LargeDocument
- *
- * 大文档性能测试 (50+ 章节)
- */
-export const LargeDocument: Story = {
-  args: {
-    items: LARGE_OUTLINE_DATA,
-    activeId: "ch-1",
-    draggable: true,
-  },
-  parameters: {
-    docs: {
-      description: {
-        story:
-          "大文档性能测试。包含 10 章 × 5 节 × 3 小节 = 180 个大纲项，测试滚动和渲染性能。",
       },
     },
   },
